wif spacing and json roundtrip
Added reading wif spacing value - ready for renderer.
In the course of this:
- enabled the json round trip to preserve spacing
- fixed saving wif to be ascii instead
changed wif save values form 1 to true where appropriate for readability/compatibility
- forced json to utf8 ascii
round tripped wif to json and back with rendering to ensure all working well. (Needs more changes to support default colors when missing)

Had to reorder some stuff to enable json to load and not complain about treddles and shafts and where contents is defined.