comments change  to configParser
Config like this so we can read the creation date embedded in comments
        self.config = RawConfigParser(comment_prefixes='/', allow_no_value=True)

but because we override comments from ;,# to / we will have to parse comments outrselves.
We did this because creation date is embedded in fiberworks files as a comment in TEXT section.

We need to deal with special case where the config file starts with a comment line. (TempoFiber)
In this case the simple approach will fail when reading from the file.
So we need to read the file into a string and use self.config.read_string() instead of self.config.read()

Also fix bug in auto option to render