Create wif from drawdown image
Adding new feature to create a wif from an image of a drawdown.
- can cleanup noisy images a bit.
- can reduce to core draft if (say) has repeats in image
- added repeats file to facilitate this but wil be used later for detecting repeats and mirrors in drafts.

Some minor changes to presenting color and in Draft when adding a warp or weft thread out of order.

So two commands added:
- pyweaving image (which creates a draft from bitmap and can turn images into drafts. needs quite a few shafts - say 40.
- pyweaving drawdown which takes an image of a drawdown and has a --core option which will remove duplicate repeats.

minor addition to reporting which files get created..

cleaned up the generators in cmd.py to factor common code