docs update
Addition of docstrings to most all of API
Added Color tests
Also fixed Color. Added as_drawn which is used in render.py when we modify the thread's Color because its either Black or White and so can't show a highlight (or shadow) if we selected shaded as an aspect of our style.